+++ src/pages/TutorialPage/Topics/02AppSettings/GeneralSetting.jsx
@@ -5,6 +5,8 @@ import Note from "../../../../components/Note/Note";
 import ArticleToc from "../../../../components/ArticleToc/ArticleToc";
 
 import general_default_folder from "../../../../assets/images/general-default_folder.png";
+import general_theme from "../../../../assets/images/general-theme.png";
+import general_theme_popup from "../../../../assets/images/general-theme_popup.png";
 import setting_others_upgradepro from "../../../../assets/images/setting-others-upgradepro.png";
 import setting_others_defaultfolder from "../../../../assets/images/setting-others-defaultfolder.png";
 import setting_others_support from "../../../../assets/images/setting-others-support.png";
@@ -12,9 +14,8 @@ import setting_others_support from "../../../../assets/images/setting-others-sup
 function GeneralSetting() {
 
     const toc = [
-        { topic: "Upgrade to Pro", id: "upgrade" },
+        { topic: "Theme", id: "changetheme" },
         { topic: "Default Data Folder", id: "defaultfolder" },
-        { topic: "Support", id: "support" },
     ];
 
     useEffect(() => {
@@ -23,14 +24,20 @@ function GeneralSetting() {
 
     return (
         <>
-            <h1>Other Setting</h1>
+            <h1>General Setting</h1>
 
             <ArticleToc topics={toc} />
 
-            <section id="upgrade">
-                <h3>Upgrade to Pro</h3>
-                <img className="no-border" src={setting_others_upgradepro} alt="Upgrade to Pro Setting" />
-                <p>Click [Upgrade to Pro] to go to the purchase page.</p>
+            <section id="changetheme">
+                <h3>Theme</h3>
+                <img className="no-border" src={general_theme} alt="Change theme" />
+                <p>Click to select the theme</p>
+                <ol className="number-list">
+                    <li>
+                        It will prompt a popup where you can choose <b>Light</b> or <b>Dark</b> Theme
+                        <img src={general_theme_popup} alt="Theme selection" />
+                    </li>
+                </ol>
             </section>
 
             <section id="defaultfolder">
@@ -46,7 +53,7 @@ function GeneralSetting() {
                 <Note>
                     <ol className="number-list">
                         <li>
-                            The [Folder Picker] interface may vary depending on your device
+                            The [Folder Picker] interface (as shown above) may vary depending on your device
                         </li>
                         <li>
                             On most device, you can create a new folder in the [Folder Picker] dialog
@@ -60,21 +67,7 @@ function GeneralSetting() {
                 </Note>
             </section>
 
-            <section id="support">
-                <h3>Support</h3>
-                <img className="no-border" src={setting_others_support} alt="Support Setting" />
-                <ol className="number-list">
-                    <li>
-                        Click [Help] to go to the online help page.
-                    </li>
-                    <li>
-                        Click [Report Issue] to go to the issue tracker or request feature page.
-                    </li>
-                    <li>
-                        Click [Contact Support] to send an email to the support team.
-                    </li>
-                </ol>
-            </section>
+            
         </>
     );
 }