+++ src/pages/TutorialPage/TutorialPage.jsx
@@ -9,6 +9,7 @@ import GeneralSetting from "./Topics/02AppSettings/GeneralSetting";
 import VaultSetting from "./Topics/02AppSettings/VaultSetting";
 import DataSetting from "./Topics/02AppSettings/DataSetting";
 import SecuritySetting from "./Topics/02AppSettings/SecuritySetting";
+import OthersSetting from "./Topics/02AppSettings/OthersSetting";
 
 // ?topic=id to link to the topic from other page
 
@@ -22,11 +23,11 @@ const tocData = [
     {
         id: "app-settings", title: "App Settings",
         subtopics: [
-            /*{ id: "general", title: "General", component: GeneralSetting },*/
+            { id: "general", title: "General", component: GeneralSetting },
             { id: "vault", title: "Vault", component: VaultSetting },
             { id: "data", title: "Data", component: DataSetting },
             { id: "security", title: "Security", component: SecuritySetting },
-            { id: "others", title: "Others", component: GeneralSetting },
+            { id: "others", title: "Others", component: OthersSetting },
         ]
     },
     {